Binary Search Tree Project
This one was really difficult, the CS modules overall have been the most challenging part of TOP so far. I have a better understanding of tree methods and recursion overall due to this project.

I am happy to have this project to refer back to when working with binary search trees going forward.